:fire: refactor cache and model loading: streamline database schema retrieval, reintroduce model loading with error handling, and enhance cache functionality
